+++ lib/entities/enums/environment.dart
@@ -1,6 +1,13 @@
 enum Environment {
+  /// Production environment
   production,
+
+  /// Staging environment (for testing)
   stage,
+
+  /// Development environment (for testing)
   development,
+
+  /// Local environment ( )
   master,
 }