+++ lib/entities/prizorsdk_params.dart
@@ -6,18 +6,47 @@ import 'package:prizorsdk_flutter/helpers/enum_helper.dart';
 import 'enums/environment.dart';
 
 class PrizorSdkParams {
+  /// Environment of the app: [production, stage, development, master]. Use [Environment.production] for production.
   final Environment? environment;
-  final String? appId;
-  final String? secret;
-  final String? appName;
+
+  /// App ID provided by CRMALL.
+  final String appId;
+
+  /// App secret provided by CRMALL.
+  final String secret;
+
+  /// App name to be shown in the app.
+  final String appName;
+
+  /// Primary color of the app, used in AppBar, buttons, icons, etc.
   final Color accentColor;
+
+  /// Foreground color of the app.
   final Color accentForegroundColor;
+
+  /// User data to be used in the app campaigns.
   final User user;
+
+  /// Image in the campaigns. If it is null, the logo will not be shown. USE PNG FORMAT.
   final Uri? appLogo;
+
+  /// Image in the home screen. If it is null, the brand will not be shown. USE PNG FORMAT.
   final Uri? appBrand;
+
+  /// Splash screen foreground color. If it is null, [accentColor] will be used.
   final Color? splashForegroundColor;
+
+  /// Splash screen background color. If it is null, [accentForegroundColor] will be used.
   final Color? splashBackgroundColor;
+
+  /// SAC data to be used in the app campaigns.
   final Sac? sac;
+
+  /// If it is true, the app will be opened in the root route Use with BottomNavigationBar, TabBar, etc.
+  final bool isRootRoute;
+
+  /// If it is true, the app will be opened in the title route Use with AppBar, etc.
+  final String titleRoute;
 
   const PrizorSdkParams({
     required this.environment,
@@ -32,6 +61,8 @@ class PrizorSdkParams {
     this.splashForegroundColor,
     this.splashBackgroundColor,
     this.sac,
+    this.isRootRoute = false,
+    this.titleRoute = '',
   });
 
   Map<String, dynamic> toJson() {
@@ -48,6 +79,8 @@ class PrizorSdkParams {
       'splashForegroundColor': splashForegroundColor?.value.toRadixString(16),
       'splashBackgroundColor': splashBackgroundColor?.value.toRadixString(16),
       'sac': sac?.toJson(),
+      'isRootRoute': isRootRoute,
+      'titleRoute': titleRoute,
     };
   }
 
@@ -65,6 +98,7 @@ class PrizorSdkParams {
       splashForegroundColor: Color(json['splashForegroundColor']),
       splashBackgroundColor: Color(json['splashBackgroundColor']),
       sac: Sac.fromJson(json['sac']),
+      isRootRoute: json['isRootRoute'],
     );
   }
 }