+++ lib/entities/user/user.dart
@@ -6,12 +6,17 @@ import 'package:prizorsdk_flutter/entities/user/phone.dart';
 import 'package:prizorsdk_flutter/helpers/enum_helper.dart';
 
 class User {
-  final String? cellphone;
-  final String? documentNumber;
+  /// [cellphone] is required because it is used to send the Authentication Code
+  final String cellphone;
+
+  /// [documentNumber] is required because it is used to Authentication and to match the user in the app
+  final String documentNumber;
   final String? name;
   final String? otherDocument;
   final Gender? gender;
   final MaritalStatus? maritalStatus;
+
+  /// Birthday: [day], [month] and [year] are required because it is used to Authentication and to match the user in the app
   final int? day;
   final int? month;
   final int? year;