+++ CHANGELOG.md
@@ -1,3 +1,20 @@
 ## 0.0.1
 
 - Initial release
+
+## 0.0.2
+
+- Updated Documentation
+- Added `Loading` Widget
+
+
+## 0.0.3
+
+- Added params `isRootPage`
+- Added params `titleRoute`
+
+## 0.0.5
+- Updated Documentation
+
+## 0.0.6
+- Changed Cep(zipcode) to String